Tristan reviewed the years of development in the develop branch to put this merge together, which brings the demo master up to date and allow us to create a release version.

+++ src/wpipe/__init__.py
@@ -121,6 +121,7 @@ __version__
     Wpipe version string
 """
 from .__metadata__ import *
+from .constants import WPIPE_NO_SCHEDULER
 from .core import *
 from .User import User
 from .Node import Node
@@ -135,45 +136,55 @@ from .Task import Task
 from .Mask import Mask
 from .Job import Job
 from .Event import Event
-from .scheduler import PbsScheduler
-from .scheduler import checkPbsConnection
+from .scheduler import PbsConsumer
+from .scheduler import SlurmConsumer
+from .scheduler.ConsumerFactory import get_consumer_factory
 
 __all__ = ['__version__', 'PARSER', 'User', 'Node', 'Pipeline', 'Input',
            'Option', 'Target', 'Configuration', 'Parameter', 'DataProduct',
-           'Task', 'Mask', 'Job', 'Event', 'PbsScheduler', 'DefaultUser',
-           'DefaultNode', 'wingspipe']
+           'Task', 'Mask', 'Job', 'Event',
+           'DefaultUser', 'DefaultNode', 'wingspipe']
+
+
+warnings.filterwarnings("ignore", message=".*Cannot correctly sort tables;.*")
 
 DefaultUser = User()
 """
 User object: User object constructed at wpipe importation (see User doc Notes)
 """
 
+PbsConsumer.DEFAULT_PORT = PbsConsumer.BASE_PORT + DefaultUser.user_id
+SlurmConsumer.DEFAULT_PORT = SlurmConsumer.BASE_PORT + DefaultUser.user_id
+
 DefaultNode = Node()
 """
 Node object: Node object constructed at wpipe importation (see Node doc Notes)
 """
 
-if PARSER.parse_known_args()[0].event_id is not None or PARSER.parse_known_args()[0].job_id is not None:
-    if PARSER.parse_known_args()[0].event_id is not None:
-        ThisEvent = Event()
-        if ThisEvent.fired_jobs[-1].is_active if len(ThisEvent.fired_jobs) else False:
-            print("Event with id %d has a job attempt that is currently running - exiting" % ThisEvent.event_id)
-            sys.exit()
-        else:
-            ThisJob = ThisEvent._generate_new_job(Task(ThisEvent.pipeline, os.path.basename(sys.argv[0])))
-            sys.argv += ['-j', str(ThisJob.job_id)]  # MEH
-    elif PARSER.parse_known_args()[0].job_id is not None:
-        ThisJob = Job()
-        ThisEvent = ThisJob.firing_event
-        if ThisJob.is_active:
-            print("Job with id %d is currently running - exiting" % ThisJob.job_id)
-            sys.exit()
-        elif not ThisJob.not_submitted:
-            ThisJob.reset()
-    ThisJob._starting_todo()
-    atexit.register(ThisJob._ending_todo)
+if pathlib.Path(sys.argv[0]).resolve().name != 'wingspipe':
+    if PARSER.parse_known_args()[0].event_id is not None or PARSER.parse_known_args()[0].job_id is not None:
+        if PARSER.parse_known_args()[0].event_id is not None:
+            ThisEvent = Event()
+            if ThisEvent.fired_jobs[-1].is_active if len(ThisEvent.fired_jobs) else False:
+                print("Event with id %d has a job attempt that is currently running - exiting" % ThisEvent.event_id)
+                sys.exit()
+            else:
+                ThisJob = ThisEvent._generate_new_job(Task(ThisEvent.pipeline, os.path.basename(sys.argv[0])))
+                sys.argv += ['-j', str(ThisJob.job_id)]  # MEH
+        elif PARSER.parse_known_args()[0].job_id is not None:
+            ThisJob = Job()
+            ThisEvent = ThisJob.firing_event
+            if ThisJob.is_active:
+                #print("Job with id %d is currently running - exiting" % ThisJob.job_id)
+                #sys.exit()
+                ThisJob.reset()
+            elif not ThisJob.not_submitted:
+                ThisJob.reset()
+        ThisJob._starting_todo()
+        atexit.register(ThisJob._ending_todo)
 
 
+# TODO: Delete?
 # def sql_hyak(task, job_id, event_id):
 #     my_job = Job(job_id)
 #     my_pipe = my_job.pipeline
@@ -239,6 +250,7 @@ if PARSER.parse_known_args()[0].event_id is not None or PARSER.parse_known_args(
 #             executable + ' -e ' + eidstr + ' -j ' + jidstr + '\n')
 #     subprocess.run(['qsub', pbsfile], cwd=my_config.confpath)
 
+
 def wingspipe(args=None):
     """
     Function that runs the wingspipe executable functionalities.
@@ -265,7 +277,9 @@ def wingspipe(args=None):
     """
     if args is not None:
         sys.argv += args  # MEH
+    # _temp = PbsConsumer.DEFAULT_PORT
     importlib.reload(sys.modules[__name__])
+    # PbsConsumer.DEFAULT_PORT = _temp
     parent_parser = si.argparse.ArgumentParser(parents=[PARSER], add_help=False)
     parser = si.argparse.ArgumentParser(prog='wingspipe', parents=[si.PARSER], add_help=False)
     subparsers = parser.add_subparsers()
@@ -285,6 +299,8 @@ def wingspipe(args=None):
     parser_run.set_defaults(which='run')
     parser_diagnose = subparsers.add_parser('diagnose', parents=[parent_parser], add_help=False)
     parser_diagnose.set_defaults(which='diagnose')
+    parser_expire = subparsers.add_parser('expire', parents=[parent_parser], add_help=False)
+    parser_expire.set_defaults(which='expire')
     parent_parser_with_yes_flag = si.argparse.ArgumentParser(parents=[parent_parser], add_help=False)
     parent_parser_with_yes_flag.add_argument('--yes', '-y', dest='yes', action='store_true',
                                              help="Don't ask for confirmation")
@@ -293,44 +309,43 @@ def wingspipe(args=None):
     parser_clean = subparsers.add_parser('clean', parents=[parent_parser_with_yes_flag], add_help=False)
     parser_clean.set_defaults(which='clean')
     parser_delete = subparsers.add_parser('delete', parents=[parent_parser_with_yes_flag], add_help=False)
+    parser_delete.add_argument('--force', '-f', dest='force', action='store_true',  # TODO
+                               help="Force deletion of every files")
     parser_delete.set_defaults(which='delete')
     args = parser.parse_args()
     if hasattr(args, 'which'):
-        my_pipe = Pipeline()
-        command = parser.prog+" "+args.which
-        if args.which == 'init':
-            my_pipe.description = args.description
-            my_pipe.attach_tasks(args.tasks_path)
-            my_pipe.attach_inputs(args.inputs_path, args.config_file)
-        elif args.which == 'run':
-            if not('WPIPE_NO_PBS_SCHEDULER' in os.environ.keys()):
-                if checkPbsConnection() != 0:
-                    print("Starting a PBS scheduler ...")
-                    import subprocess
-                    # subprocess.Popen(["python", "-m", "wpipe.scheduler.PbsConsumer"],
-                                     # stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-                    subprocess.Popen(["python", "-m", "wpipe.scheduler.PbsConsumer"]) # Have everything print to the main terminal for now
-                    time.sleep(1)
-                else:
-                    print("PBS scheduler already running ...")
-            my_pipe.run()
+        if args.which == 'expire':
+            Job(args.job_id).expire()
+        else:
+            my_pipe = Pipeline()
+            command = parser.prog + " " + args.which
+            if args.which == 'init':
+                my_pipe.description = args.description
+                my_pipe.attach_tasks(args.tasks_path)
+                my_pipe.attach_inputs(args.inputs_path, args.config_file)
+            elif args.which == 'run':
+                if not WPIPE_NO_SCHEDULER:
+                    consumer = get_consumer_factory()
+                    consumer('start')
+                # TODO if args.event_id or args.job_id
+                my_pipe.run()
 
-        elif args.which == 'diagnose':
-            my_pipe.diagnose()
-        elif args.which == 'reset':
-            if True if args.yes \
-                    else input(command + ': confirm reset of pipeline at ' +
-                               my_pipe.pipe_root + '? [y/yes] ') in ['y', 'yes']:
-                my_pipe.reset()
-        elif args.which == 'clean':
-            if True if args.yes \
-                    else input(command + ': confirm clean-up of pipeline at ' +
-                               my_pipe.pipe_root + '? [y/yes] ') in ['y', 'yes']:
-                my_pipe.clean()
-        elif args.which == 'delete':
-            if True if args.yes \
-                    else input(command + ': confirm deletion of pipeline at ' +
-                               my_pipe.pipe_root + '? [y/yes] ') in ['y', 'yes']:
-                my_pipe.delete()
+            elif args.which == 'diagnose':
+                my_pipe.diagnose()
+            elif args.which == 'reset':
+                if True if args.yes \
+                        else input(command + ': confirm reset of pipeline at ' +
+                                   my_pipe.pipe_root + '? [y/yes] ') in ['y', 'yes']:
+                    my_pipe.reset()
+            elif args.which == 'clean':
+                if True if args.yes \
+                        else input(command + ': confirm clean-up of pipeline at ' +
+                                   my_pipe.pipe_root + '? [y/yes] ') in ['y', 'yes']:
+                    my_pipe.clean()
+            elif args.which == 'delete':
+                if True if args.yes \
+                        else input(command + ': confirm deletion of pipeline at ' +
+                                   my_pipe.pipe_root + '? [y/yes] ') in ['y', 'yes']:
+                    my_pipe.delete()  # TODO: my_pipe.delete(force = args.force)
     else:
         parser.print_help()
